requirements.txt
anthropic
voyageai
PyPDF2
langchain-core
langchain-text-splitters
python-dotenv
chromadb
